#10 - Refatoração do repositório.
